[v0.91.4][maintainability] Split adl/src/cli/pr_cmd/doctor.rs
